1.2.2 1、优化自动上传触发逻辑 2、自动上传可自动触发装卡

+++ src/main/java/com/gali/ae2_auto_pattern_upload/mixin/ae2fc/ContainerFluidPatternTerminalExMixin.java
@@ -74,7 +74,7 @@ public abstract class ContainerFluidPatternTerminalExMixin {
 
             if (extracted != null) {
                 if (blanks != null) {
-                    blanks.stackSize += extracted.getStackSize();
+                    blanks.stackSize += (int) extracted.getStackSize();
                 } else {
                     blanks = extracted.getItemStack();
                 }
